tidy(nodes): move batch nodes to own file
